requirements.txt
flask
gunicorn
pymysql
python-dotenv
sqlalchemy